moved game Magic Cards to the games package
