+++ gazeplay/src/main/java/net/gazeplay/games/magiccards/MagicCardsGamesStats.java
@@ -1,6 +1,7 @@
-package utils.games.stats;
+package net.gazeplay.games.magiccards;
 
 import javafx.scene.Scene;
+import utils.games.stats.HiddenItemsGamesStats;
 
 public class MagicCardsGamesStats extends HiddenItemsGamesStats {
